Change type of variable

+++ src/BmiService.java
@@ -2,8 +2,8 @@ public class BmiService {
 
     //ИМТ = масса / рост^2//
 
-    public double calculate(int weightKilo, double heightMetres) {
-        double bmi = weightKilo / (heightMetres * heightMetres);
+    public int calculate(int weightKilo, double heightMetres) {
+        int bmi = (int) (weightKilo / (heightMetres * heightMetres));
         return bmi;
 
     }
